+++ default.nix
@@ -1,10 +1,5 @@
-# Example: Custom Holochain And Binaries
-#
-# The following `shell.nix` file can be used in your project's root folder and activated with `nix-shell`.
-# It uses a custom revision and a custom set of binaries to be installed.
-
 {
-  holonixPath ?  builtins.fetchTarball { url = "https://github.com/holochain/holonix/archive/a0dcdfac2c8783c58805175dd5bc5528ccbb35fd.tar.gz"; }
+  holonixPath ?  builtins.fetchTarball { url = "https://github.com/holochain/holonix/archive/48a75e79b1713334ab0086767a214e5b1619d38d.tar.gz"; }
 }:
 
 let
@@ -17,17 +12,18 @@ let
     holochainVersionId = "custom";
 
     holochainVersion = {
-      rev = "4347e4dbfd4f957576275fb2e8b3deda90ccbfd7";
-      sha256 = "sha256:1g3sm1x786zr9w324kxlsf50ajrmpigjj6l1xnm1cwl2hbqq7hxz";
-      cargoSha256 = "sha256:1i6i80vf7jjw1h0b3dsh5n0x8g5g3h16sw9rskw84yipqbv51nc7";
+      rev = "holochain-0.0.109";
+      sha256 = "1rwss1y8cd52ccd0875pfpbw6v518vcry3hjc1lja69x2g2x12qb";
+      cargoSha256 = "08a72d7nqpakml657z9vla739cbg8y046av4pwisdgj1ykyzyi60";
       bins = {
         holochain = "holochain";
         hc = "hc";
+        kitsune-p2p-proxy = "kitsune_p2p/proxy";
       };
 
       lairKeystoreHashes = {
-        sha256 = "1ibynj1mn1mc59x7b2jn8l1vv9m8czwcvpq81qgbpa52jgjqlf14";
-        cargoSha256 = "1dnfjdk3b4l7ysvm81r061mxly889bbcmg2h11nkgmfj79djka9s";
+        sha256 = "12n1h94b1r410lbdg4waj5jsx3rafscnw5qnhz3ky98lkdc1mnl3";
+        cargoSha256 = "0axr1b2hc0hhik0vrs6sm412cfndk358grfnax9wv4vdpm8bq33m";
       };
     };
   };
