+++ dna/zomes/profiles/Cargo.toml
@@ -10,6 +10,3 @@ name = "profiles"
 
 [dependencies]
 hc_zome_profiles = {git = "https://github.com/holochain-open-dev/profiles", branch = "main", package = "hc_zome_profiles"}
-
-hdk = "0.0.107"
-holo_hash = "0.0.7"
