--- ui/apps/who/.gitignore
@@ -1,3 +1,0 @@
-dist
-.dist
-ui.zip
