--- ui/apps/who/index.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html lang="en-GB">
-  <head>
-    <meta charset="utf-8" />
-    <style>
-      body {
-        background: #fafafa;
-      }
-    </style>
-  </head>
-  <body>
-    <who-app></who-app>
-
-    <script type="module">
-      import { WhoApp } from './src/index.ts';
-      customElements.define('who-app', WhoApp);
-    </script>
-  </body>
-</html>
