--- ui/apps/who/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "who",
-  "version": "0.0.1",
-  "scripts": {
-    "build": "rimraf dist && rollup --config rollup.config.js",
-    "build:watch": "rollup -c -w",
-    "start": "npm run build && concurrently -k --names tsc,dev-server \"npm run build:watch\" \"web-dev-server --config ./web-dev-server.config.mjs\"",
-    "package": "rimraf ui.zip && npm run build && cd ./dist && bestzip ../ui.zip * "
-  },
-  "devDependencies": {
-    "@babel/preset-env": "^7.15.0",
-    "@rollup/plugin-babel": "^5.3.0",
-    "@rollup/plugin-commonjs": "^20.0.0",
-    "@rollup/plugin-node-resolve": "^13.0.4",
-    "@rollup/plugin-replace": "^3.0.0",
-    "@rollup/plugin-typescript": "^8.2.5",
-    "@web/dev-server": "^0.1.21",
-    "@web/rollup-plugin-html": "^1.9.1",
-    "@web/rollup-plugin-import-meta-assets": "^1.0.7",
-    "babel-plugin-template-html-minifier": "^4.1.0",
-    "bestzip": "^2.2.0",
-    "concurrently": "^5.3.0",
-    "deepmerge": "^4.2.2",
-    "rimraf": "^3.0.2",
-    "rollup": "^2.56.2",
-    "rollup-plugin-node-builtins": "^2.1.2",
-    "rollup-plugin-node-globals": "^1.4.0",
-    "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-workbox": "^6.2.0",
-    "tslib": "^2.3.1",
-    "typescript": "4.3.5"
-  },
-  "dependencies": {
-    "@holochain-open-dev/cell-client": "^0.0.4",
-    "@holochain-open-dev/profiles": "github:holochain-open-dev/profiles#ui-build",
-    "@holochain/conductor-api": "^0.2.1",
-    "@lit-labs/context": "github:holochain-open-dev/context#build",
-    "@who/elements": "^0.0.1",
-    "lit": "^2.0.0-rc.4"
-  }
-}
