--- ui/apps/who/src/index.ts
@@ -1,1 +1,0 @@
-export * from './who-app';
