--- ui/apps/who/web-dev.plugins.mjs
@@ -1,16 +1,0 @@
-import { fromRollup } from '@web/dev-server-rollup';
-import rollupCommonjs from '@rollup/plugin-commonjs';
-import rollupBuiltins from 'rollup-plugin-node-builtins';
-import rollupGlobals from 'rollup-plugin-node-globals';
-import rollupReplace from '@rollup/plugin-replace';
-
-const replace = fromRollup(rollupReplace);
-const builtins = fromRollup(rollupBuiltins);
-const commonjs = fromRollup(rollupCommonjs);
-const globals = fromRollup(rollupGlobals);
-
-export default [
-  replace({
-    'process.env.PORT': process.env.PORT,
-  }),
-];
