--- ui/lib/src/index.ts
@@ -1,9 +1,0 @@
-// Import the container media queries mixin
-import 'cqfill';
-
-// TODO: change exports to be available for consumer packages
-
-export * from './elements/who-controller';
-export * from './who.service';
-export * from './who.store';
-export * from './types';
