--- ui/lib/src/sharedStyles.ts
@@ -1,12 +1,0 @@
-import { css } from 'lit';
-
-export const sharedStyles = css`
-  .column {
-    display: flex;
-    flex-direction: column;
-  }
-  .row {
-    display: flex;
-    flex-direction: row;
-  }
-`;
