--- ui/lib/src/who.service.ts
@@ -1,23 +1,0 @@
-import { CellClient } from '@holochain-open-dev/cell-client';
-import { HoloHashed, serializeHash, EntryHashB64, HeaderHashB64, AgentPubKeyB64 } from '@holochain-open-dev/core-types';
-import { Signal,} from './types';
-
-
-export class WhoService {
-  constructor(
-    public cellClient: CellClient,
-    protected zomeName = 'hc_zome_who'
-  ) {}
-
-  get myAgentPubKey() : AgentPubKeyB64 {
-    return serializeHash(this.cellClient.cellId[1]);
-  }
-
-  async notify(signal: Signal, folks: Array<AgentPubKeyB64>): Promise<void> {
-    return this.callZome('notify', {signal, folks});
-  }
-
-  private callZome(fn_name: string, payload: any) {
-    return this.cellClient.callZome(this.zomeName, fn_name, payload);
-  }
-}
