--- ui/lib/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "declarationDir": "dist",
-    "rootDir": "./src",
-    "declaration": true
-  },
-
-  "include": ["src/**/*.ts"]
-}
