+++ ui/rollup.config.js
@@ -11,7 +11,7 @@ export default {
   input: `src/index.ts`,
   output: [{ dir: 'dist', format: 'es', sourcemap: true }],
   // Indicate here external modules you don't wanna include in your bundle (i.e.: 'lodash-es')
-  external: [...Object.keys(pkg.dependencies), /lit/],
+  external: [],
   watch: {
     include: 'src/**',
     clearScreen: false,
