+++ ui/tsconfig.json
@@ -1,9 +1,11 @@
 {
-  "extends": "../../../tsconfig.json",
+  "extends": "../tsconfig.json",
   "compilerOptions": {
     "outDir": "dist",
+    "declarationDir": "dist",
     "rootDir": "./src",
-    "declaration": false
+    "declaration": true
   },
+
   "include": ["src/**/*.ts"]
 }
